Add files via upload
This file will extract part prices from the following webpage and store the values in a csv file: 
https://ca.pcpartpicker.com/guide/FkrxFT/excellent-amd-gamingstreaming-build

By running this python script repeatedly you can generate a dataset of part prices that change over time, as that webpage is updated with current prices on a daily basis. Note: the webpage is subject to change, leading to this code becoming obsolete. If this occurs small revisions must be made to the code in order to continue remaining compatible with pcpartpicker.com's build guides.